requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
mysqlclient==2.0.3
pkg_resources==0.0.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
